# web/openapi.yaml
openapi: 3.0.3
info:
  title: work-tracker API
  description: Automatically generated OpenAPI spec from Postman collection "work-tracker"
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Default local server
paths:
...
  /api/holidays:
    get:
      summary: List a year's holiday calendar
      description: The effective entry per date (a manual entry wins over the government one).
      operationId: listHolidays
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: string
            example: '2026'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidaysResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/holidays/sync:
    post:
      summary: Sync the government office calendar now (admin)
      description: Replaces the government entries of this year and the next; manual entries are kept.
      operationId: syncHolidays
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncHolidaysResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '502':
          description: The office calendar could not be fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/holidays/{date}:
    parameters:
      - name: date
        in: path
        required: true
        schema:
          type: string
          format: date
    put:
      summary: Set a manual holiday or makeup workday (admin)
      description: Creates or replaces the manual entry of the date; it wins over the government entry.
      operationId: saveHoliday
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveHolidayRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a manual holiday entry (admin)
      description: The government entry of the date, if any, applies again.
      operationId: deleteHoliday
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
  schemas:
...
    HolidayType:
      type: string
      description: holiday = a weekday off; workday = a working weekend day.
      enum:
        - holiday
        - workday
    HolidaySource:
      type: string
      description: gov = synced from the office calendar; manual = set by an admin (wins over gov).
      enum:
        - gov
        - manual
    Holiday:
      type: object
      required:
        - date
        - name
        - type
        - source
      properties:
        date:
          type: string
          format: date
        name:
          type: string
        type:
          $ref: '#/components/schemas/HolidayType'
        source:
          $ref: '#/components/schemas/HolidaySource'
    HolidaysResponse:
      type: object
      required:
        - message
        - holidays
      properties:
        message:
          type: string
        holidays:
          type: array
          items:
            $ref: '#/components/schemas/Holiday'
        lastSyncedAt:
          type: string
          format: date-time
          description: Missing before the first successful sync.
    SaveHolidayRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/HolidayType'
    HolidayResponse:
      type: object
      required:
        - message
        - holiday
      properties:
        message:
          type: string
        holiday:
          $ref: '#/components/schemas/Holiday'
    SyncHolidaysResponse:
      type: object
      required:
        - message
        - synced
        - lastSyncedAt
      properties:
        message:
          type: string
        synced:
          type: integer
        lastSyncedAt:
          type: string
          format: date-time
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string